adding mongo to refactored config
