+++ pkg/server/mongo/mongo.go
@@ -5,10 +5,11 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/4chain-ag/go-overlay-services/pkg/server/config"
 	"github.com/gookit/slog"
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
+
+	"github.com/4chain-ag/go-overlay-services/pkg/config"
 )
 
 // Client represents the MongoDB client and database connection
